pyth/requirements.txt
fastapi==0.73.0
uvicorn[standard]==0.17.1
google-cloud-speech

